Trim btlVoice 002e9450 pointer returns

+++ src/Event/mt_evtMain.c
@@ -7293,7 +7293,7 @@ u32 FUN_00367190(int param_1,int param_2,int param_3)
 // FUN_00367230 NONMATCHING
 
 
-u64 FUN_00367230(u64 param_1,u64 param_2,int param_3)
+u64 FUN_00367230(u64 param_1,u64 param_2,u64 param_3)
